chatbot/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
chainlit>=1.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pyjwt[crypto]>=2.9.0
plotly>=5.24.0
python-dotenv>=1.0.0
httpx>=0.27.0
supervisor>=4.2.0
sentence-transformers>=2.2.0
Pillow>=10.0.0
